% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_sede_puestos_control.R
\name{cp_sede_puestos_control}
\alias{cp_sede_puestos_control}
\title{Get Registry of Control Posts and Headquarters}
\source{
SERFOR – Servicio Nacional Forestal y de Fauna Silvestre.
System: SNIFFS – Sistema Nacional de Información Forestal y de Fauna Silvestre.
}
\usage{
cp_sede_puestos_control()
}
\value{
A tibble containing the registry of control posts and headquarters, with cleaned column names.
}
\description{
Retrieves data from the "Sedes y Puestos de Control" (Headquarters and Control Posts)
section under the "Control de Productos Forestales" component of the
National Forest and Wildlife Information System (SNIFFS), administered by SERFOR.

This function downloads and processes official data from the Peruvian National Forest
and Wildlife Service (SERFOR) through a shared Google Sheet.
}
\details{
The dataset includes information related to control points for the monitoring and enforcement
of forest and wildlife products in Peru.
}
\note{
Requires an active internet connection. Data is hosted on Google Drive and updated periodically by SERFOR.
}
\author{
Paul Efren Santos Andrade
}
