% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_declaraciones_manejo.R
\name{css_declaraciones_manejo}
\alias{css_declaraciones_manejo}
\title{Get Registry of Bamboo Plantations}
\source{
SERFOR (National Forest and Wildlife Service) — SNIFFS database
}
\usage{
css_declaraciones_manejo()
}
\value{
A tibble containing the cleaned dataset.
}
\description{
Downloads and processes the dataset titled "PLANES DE MANEJO Y DECLARACIONES DE MANEJO VIGENTES CAMÉLIDOS SUDAMERICANOS SILVESTRES",
available under the section \emph{CAMÉLIDOS SUDAMERICANOS SILVESTRES} of the National Forest
and Wildlife Information System (SNIFFS) of Peru.

The data is officially published by SERFOR (National Forest and Wildlife Service)
and provides information about registered bamboo plantations across the country.
}
\note{
An active internet connection is required to run this function. The data source is maintained
by SERFOR and may be updated periodically.
}
\author{
Paul Efren Santos Andrade
}
