% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icrg_acceso_recursos_geneticos.R
\name{icrg_acceso_recursos_geneticos}
\alias{icrg_acceso_recursos_geneticos}
\title{Get Registry of Access Contracts to Genetic Resources}
\source{
SERFOR - National Forest and Wildlife Service. Data accessed via Google Drive.
Component: SNIFFS - National Forest and Wildlife Information System.
}
\usage{
icrg_acceso_recursos_geneticos()
}
\value{
A tibble (data frame) containing the registry data.
}
\description{
This function accesses the registry of contracts for access to genetic resources of wild flora and fauna species,
under the item "Scientific Research and Genetic Resources" from the statistical component of the
National Forest and Wildlife Information System (SNIFFS).

It downloads and processes official data from SERFOR (National Forest and Wildlife Service)
available through Google Drive.
}
\note{
Requires an active internet connection. Data may be updated periodically by SERFOR.
}
\examples{
\dontrun{
data <- icrg_acceso_recursos_geneticos()
head(data)
View(data)
}

}
\author{
Paul Efren Santos Andrade
}
