% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pic_movilizacion_bambu.R
\name{pic_movilizacion_bambu}
\alias{pic_movilizacion_bambu}
\title{Get Bamboo Mobilization Report in Peru}
\source{
SERFOR (National Forest and Wildlife Service), SNIFFS System
Google Drive link: \url{https://docs.google.com/spreadsheets/d/1-VMvpGlXezG0YMjA_U8fMizTYcAC8QKy/edit?gid=446613698#gid=446613698}
}
\usage{
pic_movilizacion_bambu()
}
\value{
A tibble containing cleaned and standardized data on bamboo mobilization.
}
\description{
This function downloads and processes the report titled
"REPORTE DE MOVILIZACIÓN DE BAMBU A NIVEL NACIONAL", available under the
PRODUCCIÓN, INDUSTRIA Y COMERCIO category from the statistical module of
the National Forest and Wildlife Information System (SNIFFS) managed by SERFOR.

The data contains detailed information about the movement of bamboo across
different departments, provinces, and districts in Peru.
}
\note{
Requires an internet connection. The file is accessed directly from a public Google Drive URL.
}
\author{
Paul Efren Santos Andrade
}
\keyword{SNIFFS}
\keyword{bamboo}
\keyword{datasets}
\keyword{forestry}
\keyword{mobilization}
